revert CohortDefinitions refactor test-utils
